ADD-router 나누기
