+++ BackEnd/app.js
@@ -4,7 +4,6 @@ const app = express();
 const http = require('http');
 const server = http.createServer(app);
 const PORT = 8000;
-const router = require('./routes');
 const session = require('express-session');
 const cors = require('cors');
 const axios = require('axios');
@@ -19,6 +18,9 @@ const {
   Useproduct,
   Volunteer,
 } = require('./model');
+
+const connectSocket = require('./socket');
+connectSocket(server);
 
 app.use('/static', express.static('static'));
 
@@ -41,13 +43,13 @@ app.use(
   })
 );
 
-const io = require('socket.io')(server, {
-  cors: {
-    origin: 'http://localhost:3000',
-  },
-});
+// const io = require('socket.io')(server, {
+//   cors: {
+//     origin: 'http://localhost:3000',
+//   },
+// });
 
-app.set('io', io);
+// app.set('io', io);
 
 // 미들웨어를 사용하여 모든 뷰에 로그인 상태(세션)를 전달
 app.use((req, res, next) => {
@@ -57,7 +59,20 @@ app.use((req, res, next) => {
   next();
 });
 
-app.use('/', router);
+// Router 부분
+// app.use('/', router);
+const userRouter = require('./routes/user');
+app.use('/user', userRouter);
+
+const studyRouter = require('./routes/study');
+app.use('/study', studyRouter);
+
+const useproductRouter = require('./routes/useproduct');
+app.use('/product', useproductRouter);
+
+// 이 부분은 통일성 있게 라우터를 작성해서 해야 한다..이 안을 바꾸면 프론트도 바꿔야 한다..
+const chatRouter = require('./routes/chatting');
+app.use('/', chatRouter);
 
 app.use(async (req, res, next) => {
   if (req.session.isAuthenticated) {
@@ -96,74 +111,6 @@ app.use(async (req, res, next) => {
     res.locals.volunteerCount = volunteerCount;
   }
   next();
-});
-
-// socket 파트
-
-// 소켓 연결시
-io.on('connection', (socket) => {
-  let user = {};
-  console.log('Socket connected:', socket.id);
-
-  // 클라이언트가 소켓 룸 생성 요청
-  socket.on('createroom', (data) => {
-    try {
-      const { r_name, userid, nickname } = data;
-      user = { r_name: r_name, userid: userid, nickname: nickname };
-
-      socket.join(r_name);
-      console.log('찍어보자...:', data);
-      io.to(r_name).emit('notice', {
-        msg: `${r_name}채팅방 생성 성공!!'`,
-      });
-    } catch (err) {
-      console.error('Room 생성 Error 발생 ', err);
-      socket.emit('error', { msg: '알수 없는 오류가 발생했습니다.' });
-    }
-  });
-
-  socket.on('joinRoom', (data) => {
-    try {
-      // 방만들기와 비슷하게 data.r_name을 써야 할지?
-      const { r_idx, r_name, nickname, userid, u_idx } = data;
-      user = {
-        r_idx: r_idx,
-        r_name: r_name,
-        nickname: nickname,
-        userid: userid,
-        u_idx: u_idx,
-      };
-      socket.join(r_name);
-      console.log('조인할려는 룸', data);
-      io.to(r_name).emit('enter', {
-        // user닉네임으로 들어올지 아니면 userid로 들어올지
-        msg: `${user.nickname} 님이 ${data.r_name} 방에 입장합니다`,
-      });
-    } catch (err) {
-      console.error(err);
-      socket.emit('error', { msg: '불러오기 실패' });
-    }
-  });
-
-  // 채팅 메시지 전송 (이 부분 조금 생각 필요?) ***********
-  socket.on('sendMsg', (data) => {
-    console.log('채팅메시지 보내기', data);
-    io.to(user.r_name).emit('chat', {
-      nickname: data.nickname,
-      msg: data.msg,
-    });
-  });
-
-  // 채팅 끊기
-  socket.on('disconnect', () => {
-    console.log('user disconnected::', user);
-    // Postman에서 테스트 해서는 알수 없다 바로 디스커넥트가 실행이 되므로
-    // 다른 사용자의 페이지에서 확인해야 한다.
-    io.emit('exit', {
-      msg: `${user.nickname}님이 방을 떠났습니다.`,
-    });
-    socket.leave(user.r_name);
-  });
 });
 
 server.listen(PORT, function () {
